10/22/2015
Added image to database - Worked out saving a blob as an image.  Only
needs to mysqli_real_escap_string() the image. no {} are needed.
base64_encode is required to show the image.  1. Not all the info on
Stackexchange is right. 2. it did point me in the right direction. 3. It
took 2 hours to figure out that it was the {} that weren't needed and
not some other encoding issue.